moved factorized Score to ScoreManager, moved ScoreManager and record to Score package
